src/inspect/requirements.txt
pandas
numpy
scipy
sockjs-tornado==1.0.3
tornado==4.4.2